index JSON documents using the REST interface

+++ backmeup-indexer-rest/src/main/java/org/backmeup/index/rest/resources/Index.java
@@ -1,9 +1,12 @@
 package org.backmeup.index.rest.resources;
 
+import java.io.IOException;
 import java.util.Set;
 
+import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
+import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
@@ -14,6 +17,7 @@ import javax.ws.rs.core.Response;
 
 import org.backmeup.index.client.IndexClient;
 import org.backmeup.index.client.IndexClientFactory;
+import org.backmeup.index.client.IndexDocument;
 import org.backmeup.index.model.FileInfo;
 import org.backmeup.index.model.FileItem;
 import org.backmeup.index.model.SearchResultAccumulator;
@@ -89,36 +93,38 @@ public class Index {
 
     @DELETE
     @Path("/{userId}")
-    public void delete( //
-            @PathParam("userId") Long userId) {
+    public Response delete( //
+            @PathParam("userId") Long userId, //
+            @QueryParam("job") Long jobId, // optional for user and timestamp 
+            @QueryParam("time") Long timestamp) { // optional
 
         try (IndexClient indexClient = getIndexClient(userId)) {
+
+            if (jobId != null && jobId != 0) {
+                indexClient.deleteRecordsForJobAndTimestamp(jobId, timestamp);
+                return status(Response.Status.ACCEPTED, "records of job deleted");
+            }
 
             indexClient.deleteRecordsForUser();
+            return status(Response.Status.ACCEPTED, "records of user deleted");
 
         }
     }
 
-    @DELETE
+    @POST
     @Path("/{userId}")
-    public void deleteForJobAndTimestamp( //
+    @Consumes(MediaType.APPLICATION_JSON)
+    public Response index( //
             @PathParam("userId") Long userId, //
-            @QueryParam("job") Long jobId, // 
-            @QueryParam("time") Long timestamp) { 
-        mandatory("job", jobId);
-        mandatory("time", timestamp);
+            IndexDocument document) throws IOException {
 
         try (IndexClient indexClient = getIndexClient(userId)) {
 
-            indexClient.deleteRecordsForJobAndTimestamp(jobId, timestamp);
+            indexClient.index(document);
+            return status(Response.Status.CREATED, "document indexed");
 
         }
     }
-
-    //PUT/POST
-    //index/
-    //        void index(IndexDocument document) throws IOException;
-    //
 
     private void mandatory(String name, String value) {
         if (value == null) {
@@ -136,6 +142,10 @@ public class Index {
         throw new WebApplicationException(Response.status(Response.Status.BAD_REQUEST). //
                 entity(name + " parameter is mandatory"). //
                 build());
+    }
+
+    private Response status(Response.Status code, String message) {
+        return Response.status(code).entity(message).build();
     }
 
 }
